requirements.txt
streamlit
openai
pyyaml
plotly
